% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dual.R
\name{balance.options}
\alias{balance.options}
\title{Function to create balance functions}
\usage{
balance.options(balance, x, target)

balance.options(balance, x, target)
}
\arguments{
\item{balance}{a list with slots "formula", "balance.functions",
"balance.mean", or "balance.sd". One of the following combinations must be provided:
\itemize{
\item "formula" and "x"
\item "balance.functions"
}}

\item{x}{The data matrix for the source population.}

\item{target}{Either: 1) a list with slots target.mean and target.sd
or a data matrix of the target population. If a data matrix is provided
then the parameter \code{balance} should include a formula.}
}
\value{
a list with slots "balance.functions", "balance.delta",
"target.mean", "target.sd".

a list with slots "balance.functions", "balance.delta",
"target.mean", "target.sd".
}
\description{
Function to create balance functions

Function to create balance functions
}
\keyword{internal}
