% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_weight.R
\name{calc_weight}
\alias{calc_weight}
\title{Estimate causal weights}
\usage{
calc_weight(
  x,
  z,
  estimand = c("ATC", "ATT", "ATE"),
  method = supported_methods(),
  options = NULL,
  weights = NULL,
  ...
)
}
\arguments{
\item{x}{A numeric matrix of covariates. You can also pass an object of class \link[=dataHolder-class]{dataHolder} or \link[=DataSim-class]{DataSim}, which will make argument \code{z} not necessary,}

\item{z}{A binary treatment indicator.}

\item{estimand}{The estimand of interest. One of "ATT","ATC", or "ATE".}

\item{method}{The method to estimate the causal weights. Must be one of the methods returned by \code{\link[=supported.methods]{supported.methods()}}.}

\item{options}{The options for the solver. Specific options depend on the solver you will be using and you can use the solver specific options functions as detailed below..}

\item{weights}{The sample weights. Should be \code{NULL} or have a weight for each observation in the data. Normalized to sum to one.}

\item{...}{Not used at this time.}
}
\value{
An object of class \link[=causalWeights-class]{causalWeights}
}
\description{
Estimate causal weights
}
\details{
We detail some of the particulars of the function arguments below.
\subsection{COT}{

This is the.main method of the package. This method relies on the
}
}
\examples{
set.seed(23483)
n <- 2^5
p <- 6
#### get data ####
data <- causalOT::Hainmueller$new(n = n, p = p)
x <- data$get_x()
z <- data$get_z()

# estimate weights
weights <- causalOT::calc_weight(x = x,
                                 z = z, 
                                 estimand = "ATE",
                                 method = "COT",
                                 options = list(lambda = 0))
#we can also use the dataSim object directly
weightsDS <- causalOT::calc_weight(x = data,
                                 z = NULL,
                                 estimand = "ATE",
                                 method = "COT",
                                 options = list(lambda = 0))
all.equal(weights@w0, weightsDS@w0)
all.equal(weights@w1, weightsDS@w1)

\dontrun{
# Needs Torch
COTweights <- calc_weight(x = data,
      z = NULL,
      estimand = "ATE",
      method = "COT",
      options = list(p = 2,
      torch.optimizer = torch::optim_lbfgs,
      niter = 2 # only two iterations so it runs quickly
      )
      )
}
}
\seealso{
\code{\link[=estimate_effect]{estimate_effect()}}
}
