% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_weight.R
\name{calc_weight_glm}
\alias{calc_weight_glm}
\title{Calculate IPW using Logistic regression}
\usage{
calc_weight_glm(
  data,
  constraint,
  estimand = c("ATE", "ATT", "ATC"),
  method = c("Logistic", "Probit"),
  ...
)
}
\arguments{
\item{data}{}

\item{constraint}{}

\item{estimand}{}

\item{...}{}
}
\value{

}
\description{
Calculate IPW using Logistic regression
}
\keyword{internal}
