% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_weight.R
\name{calc_weight_NNM}
\alias{calc_weight_NNM}
\title{Calculate Nearest Neighbor Matching weights}
\usage{
calc_weight_NNM(
  data,
  estimand = c("ATE", "ATT", "ATC", "cATE"),
  transport.matrix = FALSE,
  sample_weight = NULL,
  ...
)
}
\arguments{
\item{data}{}

\item{estimand}{}

\item{transport.matrix}{Should we calculate the transport matrix (TRUE/FALSE)? Default is FALSE.}

\item{sample_weight}{The sample weights}

\item{...}{}
}
\value{
a list object returned to main \link{calc_weight} function
}
\description{
Calculate Nearest Neighbor Matching weights
}
\keyword{internal}
