% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causalOT-package.R
\docType{package}
\name{causalOT}
\alias{causalOT}
\alias{causalOT-package}
\title{An R package to perform causal inference using optimal transport distances.}
\description{
R code to perform causal inference weighting using a variety of methods and optimizers. The code can estimate weights, estimate treatment effects, and also give variance estimates.
}
\author{
Eric Dunipace
}
