% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R, R/weightsClass.R
\docType{class}
\name{plot.causalWeights}
\alias{plot.causalWeights}
\alias{causalWeights-class}
\title{causalWeights class}
\usage{
\method{plot}{causalWeights}(
  object,
  r_eff = NULL,
  penalty,
  p = 2,
  cost = NULL,
  debias = TRUE,
  online.cost = "auto",
  diameter = NULL,
  niter = 1000,
  tol = 1e-07,
  ...
)
}
\arguments{
\item{object}{an object of class \link[=causalWeights-class]{causalWeights}}

\item{r_eff}{The r_eff used in the PSIS calculation. See \code{\link[=PSIS_diag]{PSIS_diag()}}}

\item{penalty}{The penalty parameter to use}

\item{p}{The power of the Lp distance to use. Overridden by argument \code{cost.}}

\item{cost}{A user supplied cost function. Should take arguments \code{x1}, \code{x2}, \code{p}.}

\item{debias}{Should debiased optimal transport distances be used.}

\item{online.cost}{Should the cost be calculated online? One of "auto","tensorized", "online".}

\item{diameter}{the diameter of the covariate space. Default is NULL.}

\item{niter}{the number of iterations to run the optimal transport distances}

\item{tol}{the tolerance for convergence for the optimal transport distances}

\item{...}{passed to \code{\link[=PSIS_diag]{PSIS_diag()}}}
}
\value{
The plot method returns an invisible object of class summary_causalWeights, also returned by the \code{\link[=summary.causalWeights]{summary.causalWeights()}} function.
}
\description{
causalWeights class
}
\details{
First calls summary on the causalWeights object. Then plots the diagnostics from that summary object.

This object is returned by the \code{calc_weight} function in this package. The slots can be accessed as any S4 object. There is no publicly accessible constructor function.
}
\section{Functions}{
\itemize{
\item \code{plot(causalWeights)}: 

}}
\section{Slots}{

\describe{
\item{\code{w0}}{A slot with the weights for the control group with \eqn{n_0} entries. Weights sum to 1.}

\item{\code{w1}}{The weights for the treated group with \eqn{n_1} entries. Weights sum to 1.}

\item{\code{estimand}}{A character denoting the estimand targeted by the weights. One of "ATT","ATC", or "ATE".}

\item{\code{method}}{A character denoting the method used to estimate the weights.}

\item{\code{penalty}}{A list or the selected penalty parameters, if relevant.}

\item{\code{data}}{The dataHolder object containing the original data.}

\item{\code{call}}{The call used to construct the weights.}
}}

