% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treatment_effect.R
\name{confint.causalEffect}
\alias{confint.causalEffect}
\title{Confidence Intervals for Causal Effects}
\usage{
\method{confint}{causalEffect}(object, parm, level = 0.95, ...)
}
\arguments{
\item{object}{An object of class \link[=causalEffect-class]{causalEffect}}

\item{parm}{Unused. Included to match forms of other \code{confint} functions}

\item{level}{Confidence level. Should be between 0 and 1. Default is 0.95.}

\item{...}{Currently unused.}
}
\value{
A list with slots "CI" giving the confidence bounds and "SD" giving estimates of the standard
error of the causal effects.
}
\description{
Confidence Intervals for Causal Effects
}
\examples{
# set-up data
set.seed(1234)
data <- Hainmueller$new()
data$gen_data()

# calculate quantities
weight <- calc_weight(data, method = "Logistic")
tx_eff <- estimate_effect(data = data, weights = weight)

# get asymptotic C.I.
confint(tx_eff, model = "lm",
    formula = list(treated = "y ~ .", control = "y ~ ."))
}
