% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dual.R
\name{control.options}
\alias{control.options}
\title{Optimization controls}
\usage{
control.options(control, method)

control.options(control, method)
}
\arguments{
\item{control}{provided arguments}

\item{method}{The optimization method}
}
\value{
A list with slots depending on the optimization method

A list with slots depending on the optimization method
}
\description{
Optimization controls

Optimization controls
}
\keyword{internal}
