% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridSearch.R
\name{cot_solve,ateClass-method}
\alias{cot_solve,ateClass-method}
\title{cot_solve method for ateClass objects}
\usage{
\S4method{cot_solve}{ateClass}(object)
}
\arguments{
\item{object}{ateClass.}
}
\value{
object of class \link[=causalWeights-class]{causalWeights}
}
\description{
cot_solve method for ateClass objects
}
\keyword{internal}
