% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridSearch.R
\name{cot_solve,gridSearch-method}
\alias{cot_solve,gridSearch-method}
\title{cot_solve for gridSearch}
\usage{
\S4method{cot_solve}{gridSearch}(object)
}
\arguments{
\item{object}{gridSearch.}
}
\value{
returns object of class \link[=causalWeights-class]{causalWeights}
}
\description{
cot_solve for gridSearch
}
\keyword{internal}
