% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoodClass.R
\name{cot_solve,likelihoodMethods-method}
\alias{cot_solve,likelihoodMethods-method}
\title{cot_solve method for likelihoodMethods}
\usage{
\S4method{cot_solve}{likelihoodMethods}(object)
}
\arguments{
\item{object}{likelihoodMethods.}
}
\value{
object of class \link[=causalWeights-class]{causalWeights}
}
\description{
cot_solve method for likelihoodMethods
}
\keyword{internal}
