% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridSearch.R
\name{data_separate,dataHolder,character-method}
\alias{data_separate,dataHolder,character-method}
\title{Title}
\usage{
\S4method{data_separate}{dataHolder,character}(data, estimand)
}
\arguments{
\item{data}{dataHolder.}

\item{estimand}{character.}
}
\description{
Title
}
\keyword{internal}
