% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridSearch.R
\name{data_separate.dataHolder}
\alias{data_separate.dataHolder}
\title{Title}
\usage{
\method{data_separate}{dataHolder}(data, estimand)
}
\arguments{
\item{data}{dataHolder.}

\item{estimand}{character.}
}
\description{
Title
}
\keyword{internal}
