% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataHolder.R
\docType{class}
\name{dataHolder-class}
\alias{dataHolder-class}
\title{dataHolder-class}
\description{
dataHolder-class
}
\section{Slots}{

\describe{
\item{\code{x}}{matrix. A matrix of confounders.}

\item{\code{z}}{integer. The treatment indicator, \eqn{z_i \in \{0,1\}}.}

\item{\code{y}}{numeric. The outcome data.}

\item{\code{n0}}{integer. The number of observations where z==0}

\item{\code{n1}}{integer. The number of observations where z==1}

\item{\code{weights}}{numeric. The empirical distribution of the full sample.}
}}

\keyword{internal}
