% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataHolder.R
\name{dataHolder,dataHolder-method}
\alias{dataHolder,dataHolder-method}
\alias{dataHolder,matrix-method}
\alias{dataHolder.DataSim}
\alias{dataHolder,ANY-method}
\alias{terms.dataHolder}
\title{dataHolder-methods}
\usage{
\S4method{dataHolder}{dataHolder}(x, z = NA_integer_, y = NA_real_)

\S4method{dataHolder}{matrix}(x, z, y = NA_real_, weights = NA_real_)

dataHolder.DataSim(x, z, y = NA_real_, weights = NA_real_)

\S4method{dataHolder}{ANY}(x, z = NA_integer_, y = NA_real_, weights = NA_real_)

\method{terms}{dataHolder}(x, ...)
}
\arguments{
\item{x}{dataHolder object constructed from a formula}

\item{...}{Not used at this time}
}
\value{
a list with the formula terms for treatment and, if present, outcome formulae.
}
\description{
dataHolder-methods

dataHolder-methods

dataHolder-methods

dataHolder-methods
}
\keyword{internal}
