% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataHolder.R
\name{dataHolder}
\alias{dataHolder}
\title{dataHolder}
\usage{
dataHolder(x, z, y = NA_real_, weights = NA_real_)
}
\arguments{
\item{x}{the covariate data}

\item{z}{the treatment indicator}

\item{y}{the outcome data}

\item{weights}{the empirical distribution of the sample}
}
\value{
an object of dataHolder class with slots "x", "z", "y", and "weights"
}
\description{
dataHolder
}
\examples{
x <- matrix(0, 100, 10)
z <- stats::rbinom(100, 1, 0.5)

# don't need to provide outcome
# function will assume each observation gets equal mass
dataHolder(x = x, z = z) 
}
