% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSimClass.R
\name{DataSim}
\alias{DataSim}
\title{R6 Data Generating Parent Class}
\value{
An \link[R6:R6Class]{R6} object
}
\description{
R6 Data Generating Parent Class

R6 Data Generating Parent Class
}
\details{
Can be used to make your own data simulation class.
Should have the same slots listed in this class at a minimum,
but you can add your own, of course. An easy way to do this is
to make your class inherit from this one. See the example.
}
\examples{
MyClass <- R6::R6Class("MyClass", 
inherit = DataSim,
public = list(),
private = list())
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DataSim-get_x}{\code{DataSim$get_x()}}
\item \href{#method-DataSim-get_y}{\code{DataSim$get_y()}}
\item \href{#method-DataSim-get_z}{\code{DataSim$get_z()}}
\item \href{#method-DataSim-get_n}{\code{DataSim$get_n()}}
\item \href{#method-DataSim-get_x1}{\code{DataSim$get_x1()}}
\item \href{#method-DataSim-get_x0}{\code{DataSim$get_x0()}}
\item \href{#method-DataSim-get_p}{\code{DataSim$get_p()}}
\item \href{#method-DataSim-get_tau}{\code{DataSim$get_tau()}}
\item \href{#method-DataSim-gen_data}{\code{DataSim$gen_data()}}
\item \href{#method-DataSim-clone}{\code{DataSim$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataSim-get_x"></a>}}
\if{latex}{\out{\hypertarget{method-DataSim-get_x}{}}}
\subsection{Method \code{get_x()}}{
Gets the covariate data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataSim$get_x()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataSim-get_y"></a>}}
\if{latex}{\out{\hypertarget{method-DataSim-get_y}{}}}
\subsection{Method \code{get_y()}}{
Gets the outcome vector
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataSim$get_y()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataSim-get_z"></a>}}
\if{latex}{\out{\hypertarget{method-DataSim-get_z}{}}}
\subsection{Method \code{get_z()}}{
Gets the treatment indicator
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataSim$get_z()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataSim-get_n"></a>}}
\if{latex}{\out{\hypertarget{method-DataSim-get_n}{}}}
\subsection{Method \code{get_n()}}{
Gets the number of observations
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataSim$get_n()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataSim-get_x1"></a>}}
\if{latex}{\out{\hypertarget{method-DataSim-get_x1}{}}}
\subsection{Method \code{get_x1()}}{
Gets the covariate data for the treated individuals
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataSim$get_x1()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataSim-get_x0"></a>}}
\if{latex}{\out{\hypertarget{method-DataSim-get_x0}{}}}
\subsection{Method \code{get_x0()}}{
Gets the covaraiate data for the control individuals
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataSim$get_x0()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataSim-get_p"></a>}}
\if{latex}{\out{\hypertarget{method-DataSim-get_p}{}}}
\subsection{Method \code{get_p()}}{
Gets the dimensionality covariate data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataSim$get_p()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataSim-get_tau"></a>}}
\if{latex}{\out{\hypertarget{method-DataSim-get_tau}{}}}
\subsection{Method \code{get_tau()}}{
Gets the individual treatment effects
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataSim$get_tau()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataSim-gen_data"></a>}}
\if{latex}{\out{\hypertarget{method-DataSim-gen_data}{}}}
\subsection{Method \code{gen_data()}}{
Generates the  data. Default is an empty function
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataSim$gen_data()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataSim-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DataSim-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataSim$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
