% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataHolder.R
\name{df2dataHolder,ANY,ANY,data.frame-method}
\alias{df2dataHolder,ANY,ANY,data.frame-method}
\title{df2dataHolder-methods}
\usage{
\S4method{df2dataHolder}{ANY,ANY,data.frame}(
  treatment.formula = NA_character_,
  outcome.formula = NA_character_,
  data,
  weights = NA_real_
)
}
\description{
df2dataHolder-methods
}
\keyword{internal}
