% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dist_2d_to_1d}
\alias{dist_2d_to_1d}
\title{Covert the 2-dimensional index to 1-dimensional index}
\usage{
dist_2d_to_1d(i, j, n, m)
}
\arguments{
\item{i}{Index of row}

\item{j}{Index of column}

\item{n}{Total number of rows}

\item{m}{Total number of columns}
}
\value{
a 1d index for easy matrix entry
}
\description{
Covert the 2-dimensional index to 1-dimensional index
}
\keyword{internal}
