% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.onLoad}
\alias{.onLoad}
\title{Python path}
\usage{
.onLoad(libname, pkgname)
}
\arguments{
\item{libname}{default args. unused.}

\item{pkgname}{default args. unused.}
}
\description{
Python path
}
\details{
sets variables with paths to python functions after installation.
}
\keyword{internal}
