% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESS.R
\name{ESS}
\alias{ESS}
\title{Effective Sample Size}
\usage{
ESS(x)
}
\arguments{
\item{x}{Either a vector of weights summing to 1 or an object of class
[causalWeights][causalWeights]}
}
\value{
Either a number denoting the effective sample size or if `x` is of class
[causalWeights][causalWeights], then returns a list of both values in the treatment
and control groups.
}
\description{
Effective Sample Size
}
\details{
Calculates the effective sample size as described by Kish (1965). 
However, this calculation has some problems and the [PSIS][PSIS] 
function should be used instead.
}
\examples{
x <- rep(1/100,100)
ESS(x)
}
