% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_x.R
\name{extract_x}
\alias{extract_x}
\title{Function to extract the x data}
\usage{
extract_x(data, ...)
}
\arguments{
\item{data}{An object of class \code{matrix}, \code{data.frame}, or class \code{DataSim}.}

\item{...}{extra arguments such as "balance.functions" and
"treatment.indicator"}
}
\value{
the x data as a list "x0" and "x1"
}
\description{
Function to extract the x data
}
\keyword{internal}
