% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridSearch.R
\docType{class}
\name{gridSearch-class}
\alias{gridSearch-class}
\title{gridSearch S4 class}
\description{
gridSearch S4 class
}
\section{Slots}{

\describe{
\item{\code{penalty_list}}{numeric.}

\item{\code{nboot}}{integer.}

\item{\code{solver}}{R6.}

\item{\code{method}}{character.}

\item{\code{estimand}}{character.}
}}

\keyword{internal}
