% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dual.R
\name{marginal.wass.options}
\alias{marginal.wass.options}
\title{Marginal constraints for wasserstein}
\usage{
marginal.wass.options(marg.wass, wass, x, target)
}
\arguments{
\item{marg.wass}{a list with at minimum slots "marginal.constraints". Can
also include "marginal.costs".}

\item{wass}{output from the
\link[=wasserstein.options]{wasserstein.options} function}

\item{x}{The source data matrix. Optional if include "marginal.costs" in
\code{marg.wass}.}

\item{target}{The target data matrix. Optional if include "marginal.costs" in
\code{marg.wass}.}
}
\value{
a list with slots "marginal.costs", "marginal.delta"
}
\description{
Marginal constraints for wasserstein
}
\keyword{internal}
