% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cotOOP.R
\name{Measure}
\alias{Measure}
\title{Set-up a measure}
\usage{
Measure(
  x,
  weights = NULL,
  probability.measure = TRUE,
  adapt = c("none", "weights", "x"),
  balance.functions = NA_real_,
  target.values = NA_real_,
  dtype = NULL,
  device = NULL
)
}
\arguments{
\item{x}{The data points}

\item{weights}{The empirical measure. If NULL, assigns equal weight to each observation}

\item{probability.measure}{Is the empirical measure a probability measure? Default is TRUE.}

\item{adapt}{Should we try to adapt the data ("x"), the weights ("weights"), or neither ("none"). Default is "none".}

\item{balance.functions}{A matrix of functions of the covariates to target for mean balance. If NULL and \code{target.values} are provided, will use the data in \code{x}.}

\item{target.values}{The targets for the balance functions. Should be the same length as columns in \code{balance.functions.}}

\item{dtype}{The torch_tensor dtype.}

\item{device}{The device to have the data on.}
}
\value{
An R6 object of class "Measure"
}
\description{
Set-up a measure
}
\examples{
x <- matrix(0, 100, 10)
m <- Measure(x = x)
}
