% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cotOOP.R
\name{oop_loss_select}
\alias{oop_loss_select}
\title{Internal function to select appropriate loss function}
\usage{
oop_loss_select(ot)
}
\arguments{
\item{ot}{an OT object}
}
\description{
Selects sinkhorn or energy distance losses depending on value
of penalty parameter
}
\keyword{internal}
