% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supported.methods.R
\name{ot.methods}
\alias{ot.methods}
\title{Supported optimal transport methods}
\usage{
ot.methods()
}
\value{
A character vector with values "NNM","Wasserstein", and "SCM".
}
\description{
Supported optimal transport methods
}
\examples{
ot.methods()
}
