% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OT.R
\name{ot_distance,matrix-method}
\alias{ot_distance,matrix-method}
\alias{ot_distance,array-method}
\alias{ot_distance,torch_tensor-method}
\title{ot_distance-methods}
\usage{
\S4method{ot_distance}{matrix}(
  x1,
  x2 = NULL,
  a = NULL,
  b = NULL,
  penalty,
  p = 2,
  cost = NULL,
  debias = TRUE,
  online.cost = "auto",
  diameter = NULL,
  niter = 1000,
  tol = 1e-07
)

\S4method{ot_distance}{array}(
  x1,
  x2 = NULL,
  a = NULL,
  b = NULL,
  penalty,
  p = 2,
  cost = NULL,
  debias = TRUE,
  online.cost = "auto",
  diameter = NULL,
  niter = 1000,
  tol = 1e-07
)

\S4method{ot_distance}{torch_tensor}(
  x1,
  x2 = NULL,
  a = NULL,
  b = NULL,
  penalty,
  p = 2,
  cost = NULL,
  debias = TRUE,
  online.cost = "auto",
  diameter = NULL,
  niter = 1000,
  tol = 1e-07
)
}
\description{
ot_distance-methods

ot_distance-methods

ot_distance-methods
}
\keyword{internal}
