% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.summary_causalWeights}
\alias{print.summary_causalWeights}
\alias{plot.summary_causalWeights}
\title{Plotting and print methods for summary_causalWeights}
\usage{
\method{print}{summary_causalWeights}(object, ...)

\method{plot}{summary_causalWeights}(object, ...)
}
\arguments{
\item{object}{an object of class "summary_causalWeights" returned by the function \code{\link[=summary.causalWeights]{summary.causalWeights()}}}

\item{...}{Not used}
}
\description{
Plotting and print methods for summary_causalWeights
}
\section{Functions}{
\itemize{
\item \code{print(summary_causalWeights)}: 

}}
