% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_data.R
\name{prep_data}
\alias{prep_data}
\title{prep_data function to prepare data for weighting and outcome estimation}
\usage{
prep_data(data, ...)
}
\arguments{
\item{data}{Either a matrix, a data.frame, or a DataSim class}

\item{...}{Additional arguments are necessary if a matrix or data.frame is given. These are:
\itemize{
\item{balance.covariates: }{ The names of the covariates to include for balancing}
\item{treatment.indicator: }{ The name of the covariate denoting the treatment indicator}
\item{outcome: }{ (Optional) The name of the covariate that denotes the outcome.}
}}
}
\value{
A list with slots "df" and "z". "df" is a data.frame with the covariates and outcome and "z" is the treatment indicator as an integer vector.
}
\description{
prep_data function to prepare data for weighting and outcome estimation
}
\keyword{internal}
