% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataHolder.R
\name{print.dataHolder}
\alias{print.dataHolder}
\title{print.dataHolder}
\usage{
\method{print}{dataHolder}(x, ...)
}
\arguments{
\item{x}{dataHolder object}

\item{...}{Not used}
}
\description{
print.dataHolder
}
