% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSIS.R
\name{PSIS.causalWeights}
\alias{PSIS.causalWeights}
\title{PSIS casualWeights class}
\usage{
PSIS.causalWeights(x, r_eff = NULL, ...)
}
\arguments{
\item{x}{object of class causalWeights}

\item{r_eff}{pass to PSIS}

\item{...}{pass to PSIS method}
}
\value{
object of class causalPSIS
}
\description{
PSIS casualWeights class
}
\keyword{internal}
