% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSIS_diagnostic.R
\name{PSIS}
\alias{PSIS}
\title{Pareto-Smoothed Importance Sampling}
\usage{
PSIS(x, r_eff = NULL, ...)
}
\arguments{
\item{x}{a vector of weights, an object of class [causalWeights][causalWeights], or a list with slots
"w0" and "w1".}

\item{r_eff}{A vector of relative effective sample size with one estimate per observation. If providing
an object of class [causalWeights][causalWeights], should be a list of vectors with one vector for each
sample. See [psis()][loo::psis] for more details.}

\item{...}{Arguments passed to the [psis()][loo::psis] function.}
}
\value{
An object of class "psis". See [psis()][loo::psis]. Will give the log of the 
smoothed weights in slot `log_weights`, and in the slot `diagnostics`, it will give 
the `pareto_k` parameter (see the [pareto-k-diagnostic][loo::pareto-k-diagnostic] page) and
the `n_eff` estimates.
}
\description{
Pareto-Smoothed Importance Sampling
}
\details{
Acts as a wrapper to the [psis()][loo::psis] function from the [loo][pkg::loo] package. It
is built to handle the data types found in this package. This method is preferred to the [ESS()][ESS]
function since the latter is prone to error but will not give good any indication that the estimates
are problematic.
}
\examples{
\dontrun{
x <- runif(100)
w <- x/sum(x)

res <- PSIS(x = w, r_eff = x)
}
}
