% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_weight.R
\docType{class}
\name{sampleWeights-class}
\alias{sampleWeights-class}
\title{sampleWeights class}
\description{
sampleWeights class
}
\section{Slots}{

\describe{
\item{\code{a}}{The sample weights for the fist group}

\item{\code{b}}{The sample weights for the second group}

\item{\code{total}}{The sample weights for the overall sample}
}}

