% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_search.R
\name{sbw_grid_search}
\alias{sbw_grid_search}
\title{Grid search for SBW}
\usage{
sbw_grid_search(
  data,
  grid = NULL,
  estimand = c("ATT", "ATC", "cATE", "ATE", "feasible"),
  n.boot = 1000,
  grid.length = 10,
  ...
)
}
\arguments{
\item{data}{Object of class \code{matrix}, \code{data.frame}, or \code{DataSim}.}

\item{grid}{Either NULL or a vector of values to check.}

\item{estimand}{One of "ATT", "ATC","cATE","ATE","feasible".}

\item{n.boot}{Number of bootstraps.}

\item{grid.length}{Number of grid values.}

\item{...}{extra arguments passed to \link{extract_x}}
}
\value{
an object of class \link[=causalWeights-class]{causalWeights}
}
\description{
Grid search for SBW
}
\keyword{internal}
