% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmClass.R
\name{scmOptions}
\alias{scmOptions}
\title{Options for the SCM Method}
\usage{
scmOptions(...)
}
\arguments{
\item{...}{Arguments passed to the \link[osqp:osqpSettings]{osqpSettings()} function which solves the problem.}
}
\value{
A list with arguments to pass to \link[osqp:osqpSettings]{osqpSettings()}
}
\description{
Options for the SCM Method
}
\details{
Options for the solver used in the optimization of the Synthetic Control Method of Abadie and Gardeazabal (2003).
}
\examples{
opts <- scmOptions()
}
