% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supported.methods.R
\name{supported.methods}
\alias{supported.methods}
\title{Supported weighting methods}
\usage{
supported.methods()
}
\value{
A character vector with values "Logistic", "Probit", "SBW", "SCM",
"CBPS",  "NNM",
"Wasserstein" or equivalently "COT",
and "None".
}
\description{
Supported weighting methods
}
\examples{
supported.methods()
}
