% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supported.methods.R
\name{supported.solvers}
\alias{supported.solvers}
\title{Supported solvers}
\usage{
supported.solvers()
}
\value{
a character vector with values "lbfgs","mosek","gurobi", and "osqp"
}
\description{
Supported solvers
}
\details{
The solvers "mosek" and "gurobi" are commercial solvers that require software licenses. "quadprog" uses the \code{osqp} R package and "lbfgs" will either use \code{pytorch} in Python or the \code{lbfgs3c} package in R, which are both free.
}
\examples{
supported.solvers()
}
