% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportedMethods.R
\name{supported_methods}
\alias{supported_methods}
\title{Supported Methods}
\usage{
supported_methods()
}
\value{
A character list with supported methods. Note "COT" is the same as "Wasserstein". We provide the second name for backwards compatibility.
}
\description{
Supported Methods
}
\examples{
supported_methods()
}
