% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataHolder.R
\name{terms.dataHolder}
\alias{terms.dataHolder}
\title{terms function for dataHolder object}
\usage{
\method{terms}{dataHolder}(x, ...)
}
\arguments{
\item{x}{dataHolder object constructed from a formula}

\item{...}{}
}
\value{
a list with the formula terms for treatment and, if present, outcome formulae.
}
\description{
terms function for dataHolder object
}
