% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barycentric_projection.R
\name{barycentric_projection}
\alias{barycentric_projection}
\title{Barycentric Projections}
\usage{
barycentric_projection(data, weight, ...)
}
\arguments{
\item{data}{Data used to perform projections. Should be a matrix, data.frame, or DataSim class.}

\item{weight}{an object of class \link[=causalWeights-class]{causalWeights} output from the \code{\link[=calc_weight]{calc_weight()}} function}

\item{...}{optional arguments such as "cost", the cost matrix, or the "estimand", "metric", or "p" if not provided by weights. Arguments "balance.covariates" and "treatment.indicator" must be provided if data is of class data.frame or matrix.}
}
\value{
a list containing the barycentric projections with slots "control", "treated", and "observed.treatment"
}
\description{
Barycentric Projections
}
\keyword{internal}
