% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_weight.R
\name{calc_weight}
\alias{calc_weight}
\title{Estimate causal weights}
\usage{
calc_weight(
  data,
  constraint = NULL,
  estimand = c("ATE", "ATT", "ATC", "cATE", "feasible"),
  method = supported.methods(),
  formula = NULL,
  transport.matrix = FALSE,
  grid.search = FALSE,
  ...
)
}
\arguments{
\item{data}{Either a matrix, a data.frame, or a DataSim class. Arguments "balance.covariates" and "treatment.indicator" must be provided in the \code{...} arguments if data is of class data.frame or matrix.}

\item{constraint}{The constraints for the weights. See details.}

\item{estimand}{The estimand of interest. One of "ATT","ATC", or "ATE".}

\item{method}{The method to estimate the causal weights. Must be one of the methods returned by \code{\link[=supported.methods]{supported.methods()}}.}

\item{formula}{The formula for creating the design matrix used in various methods. See details.}

\item{transport.matrix}{Should the method calculate the transportation matrix if not done as a part of the method (TRUE/FALSE)? Default is FALSE.}

\item{grid.search}{Should hyperparameters be selected by a grid search? Only available for "SBW" and "Wasserstein" methods.}

\item{...}{Many additional arguments are possible depending on the chosen method. See details for more information. Arguments "balance.covariates" and "treatment.indicator" must be provided if data is of class data.frame or matrix.}
}
\value{
An object of class \link[=causalWeights-class]{causalWeights}
}
\description{
Estimate causal weights
}
\details{
#Data
The following classes are recognized by the \code{data} variable.
\subsection{DataSim class}{

The DataSim class is provided by this package for simulations. You can pass a DataSim object (once data has been simulated) to this function
and it will be recognized and handled appropriately.
}

\subsection{data.frame or matrix}{

If the \code{data} argument is of class \code{data.frame} or \code{matrix}, then additional arguments are necessary to pass in the dots (\code{...}).
These \emph{must} include a vector argument \code{balance.covariates} and an integer or character in the \code{treatment.indicator} argument. The
\code{balance.covariates} argument should be either an integer vector giving the column numbers of the covariates to balance or a character vector
giving the names of the columns to balance. Similarly, the \code{treatment.indicator} argument should be a integer giving the column number of the
treatment labels or a character giving the column name.

#Constraints
The constraint argument is used by the balancing methods like "SBW" dnd "Wasserstein" with balancing functions.
}
}
\examples{
set.seed(23483)
n <- 2^7
p <- 6
overlap <- "low"
design <- "A"
metric <- c("Lp")
power <- 1
estimate <- "ATE"
#### get simulation functions ####
data <- causalOT::Hainmueller$new(n = n, p = p, 
      design = design, overlap = overlap)
      data$gen_data()
      weights <- calc_weight(data = data, 
      p = p,
      estimand = estimate,
      method = "NNM")
}
