\name{causalOT-package}
\alias{causalOT-package}
\alias{causalOT}
\docType{package}
\title{
\packageTitle{causalOT}
}
\description{
\packageDescription{causalOT}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{causalOT}
\packageIndices{causalOT}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{causalOT}

Maintainer: \packageMaintainer{causalOT}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
