% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightsClass.R
\docType{class}
\name{causalWeights-class}
\alias{causalWeights-class}
\title{causalWeights class}
\description{
causalWeights class
}
\details{
This object is returned by the \code{calc_weight} function in this package. The slots can be accessed as any S4 object. There is no publicly accessible constructor function.
}
\section{Slots}{

\describe{
\item{\code{w0}}{A slot with the weights for the control group with \eqn{n_0} entries. Weights sum to 1.}

\item{\code{w1}}{The weights for the treated group with \eqn{n_1} entries. Weights sum to 1.}

\item{\code{estimand}}{A character denoting the estimand targeted by the weights. One of "ATT","ATC", or "ATE".}

\item{\code{info}}{A slot to store a variety of info for inference. Currently under development.}

\item{\code{method}}{A character denoting the method used to estimate the weights.}

\item{\code{penalty}}{A list or the selected penalty parameters, if relevant.}

\item{\code{data}}{The dataHolder object containing the original data.}

\item{\code{call}}{The call used to construct the weights.}
}}

