% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_weight.R
\docType{class}
\name{causalWeights-class}
\alias{causalWeights-class}
\title{causalWeights class}
\description{
causalWeights class
}
\section{Slots}{

\describe{
\item{\code{w0}}{A slot with the weights for the control group.}

\item{\code{w1}}{The weights for the treated group.}

\item{\code{gamma}}{The trasportation matrix. If estimand is "ATE", will be a list with the transportation plan for each treatment group to balance towards the overall treatment.}

\item{\code{estimand}}{A character denoting the estimand targeted by the weights. One of "ATT","ATC", or "ATE".}

\item{\code{method}}{A character denoting the method used to estimate the weights.}

\item{\code{args}}{The other arguments used to construct the weights.}
}}

