% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESS.R
\docType{methods}
\name{ESS.default}
\alias{ESS.default}
\alias{ESS.causalWeights}
\alias{ESS}
\title{Effective Sample Size}
\usage{
\method{ESS}{default}(x)

\method{ESS}{causalWeights}(x)

ESS(x)
}
\arguments{
\item{x}{Either a vector of weights summing to 1 or an object of class
\href{causalOT::causalWeights}{causalWeights}}
}
\value{
Either a number denoting the effective sample size or if \code{x} is of class
\url{causalWeights}, then returns a list of both values in the treatment
and control groups.
}
\description{
Effective Sample Size
}
\details{
Calculates the effective sample size as described by Kish (1965).
However, this calculation has some problems and the \url{PSIS}
function should be used instead.
}
\section{Functions}{
\itemize{
\item \code{ESS.default}: default ESS method for numeric vectors

\item \code{ESS.causalWeights}: ESS method for objects of class causalWeights
}}

\examples{
x <- rep(1/100,100)
ESS(x)
}
\seealso{
\href{PSIS}{PSIS()}
}
