% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cotOOP.R
\name{OTProblem}
\alias{OTProblem}
\title{Object Oriented OT Problem}
\usage{
OTProblem(measure_1, measure_2, ...)
}
\arguments{
\item{measure_1}{An object of class \link{Measure}}

\item{measure_2}{An object of class \link{Measure}}

\item{...}{Not used at this time}
}
\value{
An R6 object of class "OTProblem"
}
\description{
Object Oriented OT Problem
}
\examples{
x <- matrix(0, 100, 10)
m1 <- Measure(x = x)

y <- matrix(0, 100, 10)
m2 <- Measure(x = y)
ot <- OTProblem(m1, m2)
}
