% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cotOOP.R
\name{OTProblem}
\alias{OTProblem}
\title{Object Oriented OT Problem}
\usage{
OTProblem(measure_1, measure_2, ...)
}
\arguments{
\item{measure_1}{An object of class \link{Measure}}

\item{measure_2}{An object of class \link{Measure}}

\item{...}{Not used at this time}
}
\value{
An R6 object of class "OTProblem"
}
\description{
Object Oriented OT Problem
}
\examples{
if (torch::torch_is_installed()) {
# setup measures
x <- matrix(1, 100, 10)
m1 <- Measure(x = x)

y <- matrix(2, 100, 10)
m2 <- Measure(x = y, adapt = "weights")

z <- matrix(3,102, 10)
m3 <- Measure(x = z)

# setup OT problems
ot1 <- OTProblem(m1, m2)
ot2 <- OTProblem(m3, m2)
ot_master <- 0.5 * ot1 + 0.5 * ot2

# solve for weights
ot_master$setup_arguments(lambda = 100)
ot_master$solve(niter = 1, torch_optim = torch::optim_rmsprop)
}
}
